main: copy config file only if output is in a different folder

+++ src/impl_simulation.hpp
@@ -54,9 +54,11 @@ bool Simulation::init(string suffix) {
     file.close();
 
     // saves a copy of the config file in the output folder
-    fs::copy(config->filename,
-             config->data["OutFolder"].as<string>() + "/" + config->filename,
-             fs::copy_options::overwrite_existing);
+    if (not fs::exists(config->data["OutFolder"].as<string>() + "/config_" + suffix + ".yaml")) {
+        fs::copy(config->filename,
+                 config->data["OutFolder"].as<string>() + "/" + fs::path(config->filename).filename().string(),
+                 fs::copy_options::overwrite_existing);
+    }
 
     DEBUG = config->data["Debug"].as<bool>();
